requirements.txt
# ============================================
# Wildlife AI Platform - Python Dependencies
# ============================================
# PyTorch is installed separately via pip with CUDA index URL (see SETUP.md)
# MegaDetector is installed via pip install megadetector (see SETUP.md)
# AWC135 is installed from git clone (see SETUP.md)

# --- Web Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# --- Database ---
sqlalchemy==2.0.36
alembic==1.14.1
aiosqlite==0.20.0

# --- Authentication ---
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# --- Image Processing ---
Pillow==11.1.0
opencv-python-headless==4.10.0.84
timm==1.0.12

# --- Data Processing ---
pandas==2.2.3
numpy==1.26.4

# --- Utilities ---
tqdm==4.67.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# --- Async Workers ---
celery[redis]==5.4.0
redis==5.2.1

# --- Testing ---
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# --- Optional: Production Database (uncomment when ready) ---
# psycopg2-binary==2.9.10
# asyncpg==0.30.0
